Fix more LGTM implicit casting alerts

+++ engine/src/main/java/org/terasology/rendering/RenderHelper.java
@@ -75,7 +75,7 @@ public final class RenderHelper {
         float timeFactor = RefractiveReflectiveBlocksNode.waveSpeed;
 
         for (int i = 0; i < OCEAN_OCTAVES; ++i) {
-            height += (smoothTriangleWave(timeToTick(days,
+            height += (float) (smoothTriangleWave(timeToTick(days,
                     timeFactor) + position.x * OCEAN_WAVE_DIRECTIONS[i].x * waveSize + position.z * OCEAN_WAVE_DIRECTIONS[i].y * waveSize) * 2.0 - 1.0) * waveIntensity;
 
             waveSize *= RefractiveReflectiveBlocksNode.waveSizeFalloff;
